fix issues after 1.54.1 changes
